@@ -1,19 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <Content Remove="wwwroot/js\site.js" />
-    <Content Remove="wwwroot/js\site.min.js" />
+    <PackageReference Include="NetCore.AutoRegisterDi" Version="1.1.0" />
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="wwwroot\AddUserDefinedFunctions.sql" />
-  </ItemGroup>
+
   <ItemGroup>
     <ProjectReference Include="..\BizLogic\BizLogic.csproj" />
     <ProjectReference Include="..\DataLayer\DataLayer.csproj" />
