@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using Autofac;
 using Autofac.Extensions.DependencyInjection;
@@ -14,7 +15,10 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
+using NetCore.AutoRegisterDi;
+using ServiceLayer.BookServices;
 using ServiceLayer.DatabaseServices.Concrete;
 
 namespace EfCoreInAction
@@ -30,7 +34,7 @@ namespace EfCoreInAction
         public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
-        public IServiceProvider ConfigureServices(IServiceCollection services)
+        public void ConfigureServices(IServiceCollection services)
         {
             var gitBranchName = DatabaseStartupHelpers.GetWwwRootPath().GetBranchName();
 
@@ -42,41 +46,27 @@ namespace EfCoreInAction
             //This makes the Git branch name available via injection
             services.AddSingleton(new AppInformation(gitBranchName));
 
-            var connection = Configuration.GetConnectionString("OverrideConnection");
-            if (connection == null)
+            var connection = Configuration.GetConnectionString("DefaultConnection");
+            if (Configuration["ENVIRONMENT"] == "Development")
             {
-                connection = Configuration.GetConnectionString("DefaultConnection");
-                if (Configuration["ENVIRONMENT"] == "Development")
-                {
-                    //if running in development mode then we alter the connection to have the branch name in it
-                    connection = connection.FormDatabaseConnection(gitBranchName);
-                }
+                //if running in development mode then we alter the connection to have the branch name in it
+                connection = connection.FormDatabaseConnection(gitBranchName);
             }
-            //Swapped to DbContext Pooling
-            services.AddDbContextPool<EfCoreContext>( //#A
-                options => options.UseSqlServer(connection, //#B
-                b => b.MigrationsAssembly("DataLayer"))); //#C
-            /************************************************************
-            #A I register my applictaion DbContext using the AddDbContextPool<T>, instead of the normal AddDbContext<T> method
-            #B Here I am using a Sql Server database, but pooling works with any database provider
-            #C Because I am using migrations in a layered architecture I need to tell the database provider which assembly the migration code is in
-             * ********************************************************/
+            services.AddDbContext<EfCoreContext>(options => options.UseSqlServer(connection,
+                b => b.MigrationsAssembly("DataLayer")));
+            services.AddControllersWithViews();
+            services.AddRazorPages();
 
-            //Add AutoFac
-            var containerBuilder = new ContainerBuilder();
-            containerBuilder.RegisterModule<ServiceLayer.Utils.MyAutoFacModule>();
-            containerBuilder.Populate(services);
-            var container = containerBuilder.Build();
-            return new AutofacServiceProvider(container);
+            //replace AutoFac not so easy to use in ASP.NET Core 3 - uses one of my libraries
+            services.RegisterAssemblyPublicNonGenericClasses(Assembly.GetAssembly(typeof(BookListDto)))
+                .Where(c => c.Name.EndsWith("Service"))
+                .AsPublicImplementedInterfaces();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, 
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, 
             ILoggerFactory loggerFactory, IHttpContextAccessor httpContextAccessor)
         {
-            //Remove the standard loggers because they slow the applictaion down
-            //loggerFactory.AddConsole(Configuration.GetSection("Logging"));
-            //loggerFactory.AddDebug();
             loggerFactory.AddProvider(new RequestTransientLogger(() => httpContextAccessor));
             if (env.IsDevelopment())
             {
@@ -85,15 +75,20 @@ namespace EfCoreInAction
             else
             {
                 app.UseExceptionHandler("/Home/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+                app.UseHsts();
             }
-
+            app.UseHttpsRedirection();
             app.UseStaticFiles();
 
-            app.UseMvc(routes =>
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapRazorPages();
             });
         }
     }
